requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.49
mysql-connector-python==8.1.0
pydantic==1.10.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv==1.0.0
